backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pymongo
certifi
drf-yasg
django-filter
PyPDF2
python-docx
textract
openai